pyproject.toml
[tool.poetry]
name = "python_docker_sample"
version = "0.1.0"
description = ""
authors = ["Iuliia_volkova2 <iuliia_volkova2@epam.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.8.6"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
